requirements.txt
matplotlib~=3.9.0
torch~=2.3.1
torchvision~=0.18.1
torcheval
tqdm~=4.66.5
joblib
astropy~=6.0.1
webdavclient3
python-dotenv~=1.0.0
numpy~=1.25.2